refactor: rename field outline theme files
