+++ packages/field-highlighter/theme/lumo/vaadin-field-highlighter.js
@@ -4,7 +4,7 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import '@vaadin/vaadin-lumo-styles/user-colors.js';
+import './vaadin-field-outline-styles.js';
 import './vaadin-user-tags.js';
-import './vaadin-field-highlighter-styles.js';
 
 export * from '../../src/vaadin-field-highlighter.js';
