+++ packages/field-highlighter/theme/lumo/vaadin-field-outline-styles.js
