+++ packages/field-highlighter/theme/material/vaadin-field-outline-styles.js
